Enhanced functionality in `display_game_library()` to enable the play button feature.
Now, when the play button is clicked, the compiler script is triggered. I've configured it to use the 'Multiplayer Tetris' JSON file as the default game. This means that whenever a user clicks the play button for any game frame (e.g., game 1, game 2, etc.), the compiler will compile the Multiplayer Tetris game.

To ensure smooth operation, I added a print statement in the compiler code to verify the existence of the game JSON file path. If the file path doesn't exist, an error message is displayed.

However, it appears that the `compile.js` script is unable to locate the specified file. I suspect there might be an issue within the `compile.js` script itself, as the GUI side seems to be functioning correctly.

I recommend the team member assigned to the script review it for any potential errors.