Commented old Classroom class, and wrote a new Classroom class which combines the login feature and the classroom content.
I do not know how to pass through the login lol. But I am not getting no interpreter errors nor warnings.

Below the Classroom class, there are two versions of the Classroom class that are commented. One is Classroom with the login, but without the barebones design of the classroom. And the other one is the Classroom vice versa. These two are the old code which were used to combine the code.